arm: fix FPU trap detection for ARMv8 AArch32
The ISS encoding for HSR.EC=7 differs between ARMv7 and ARMv8.
On ARMv8 HSR.EC=7 is used solely for FPU exceptions, therefore
the coproc is defined as RES0, while on ARMv7 it is defined as
"the number of the coprocessor accessed by the trapped operation".

Our FPU trap detection assumed the ARMv7 encoding, and thus failed to
recognize certain FPU traps on ARMv8 Aarch32. Since HSR/ESR.EC=7 is
used solely for FPU exceptions on all ARMv8 implementations, we now
unconditionally treat it as such on ARMv8 Aarch32, just as we already
do for Aarch64.

Change-Id: I8c41a41145d162b3b4b6d353e025de1c519b9d96
Signed-off-by: Adam Lackorzynski <adam@l4re.org>