Fix serial input setup on x86 with interrupt remapping
So far, serial input via the kernel UART was set up during the static
initialization phase on the boot stack. On x86, with an IOMMU that
caches not-present entries, the kernel fails to invalidate the IOMMU
interrupt entry cache, after it enabled the UART's receive IRQ, since
Intel_iommu::set_irq_mapping() does not support being called on the
boot stack, as it uses Inv_desc::wait_pmem().

Thus, we move serial input setup to Kernel_thread::bootstrap(), which
runs on the kernel thread stack. This should not cause any noticeable
difference, as the kernel only opens interrupts for the first time
later in Kernel_thread::bootstrap().

The bug was observed with QEMU when the `caching-mode=on` option of the
intel-iommu QEMU device was set. Hardware platforms, however, should not
be affected, because the Intel VT-d specification mandates that hardware
implementations must not cache non-present or erroneous IRTEs.

Change-Id: Icadc2abad2f73e2b26811a01f621ea427d1ba4a3
Signed-off-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>