Arm/Gic_redist: Prevent overrunning in scan_range
GIC Redistributor ranges can be given in multiple regions. Regions are
mapped in small pages. And they probably could be distributed in
physical memory (or maybe not but specifying multiple regions make
this at least theoretically possible). Because we map and scan for each
individual core, we must also consider the size of the region we mapped,
not only the end-marker of the redistributor.

Change-Id: Iba18ae95622ab536ace128da68376bf32f1ebfc6
Signed-off-by: Adam Lackorzynski <adam@l4re.org>